Split into logwindows. Working on fixing day rollover in partitions
